+++ routes/events.js
@@ -3,7 +3,6 @@ const events = require('../models').events;
 const rooms = require('../models').rooms;
 const companies = require('../models').companies;
 
-//select * from events
 routes.get('/', (req, res) => {
     events.findAll({ include: [ { model: rooms, include: companies } ] })
         .then(events => {
@@ -25,37 +24,43 @@ routes.get('/:id', (req, res) => {
 });
 
 routes.post('/', (req, res) => {
-    events.create(req.body)
-        .then(event => {
-            res.status(201).send(event);
-        })
-        .catch(err => {
-            res.status(501).send({ message: err.message });
-        });
+    if(req.user.role === 1 || req.user.role === 2) {
+        events.create(req.body)
+            .then(event => {
+                res.status(201).send(event);
+            })
+            .catch(err => {
+                res.status(501).send({message: err.message});
+            });
+    } else res.status(500).send({ message: 'You have no rights' });
 });
 
 routes.put('/:id', (req, res) => {
-    events.findOne({ where: { id: req.params.id } })
-        .then(event => {
-            if(event) return events.update(req.body, { where : { id: req.params.id } });
-            else res.status(500).send({ message: 'Wrong id' });
-        })
-        .then(event => {
-            res.status(200).send(event);
-        })
-        .catch(err => {
-            res.status(500).send({ message: err.message });
-        });
+    if(req.user.role === 1 || req.user.role === 2) {
+        events.findOne({where: {id: req.params.id}})
+            .then(event => {
+                if (event) return events.update(req.body, {where: {id: req.params.id}});
+                else res.status(500).send({message: 'Wrong id'});
+            })
+            .then(event => {
+                res.status(200).send(event);
+            })
+            .catch(err => {
+                res.status(500).send({message: err.message});
+            });
+    } else res.status(500).send({ message: 'You have no rights' });
 });
 
 routes.delete('/:id', (req, res) => {
-    events.destroy({ where: { id: req.params.id } })
-        .then(event => {
-            event ? res.status(200).send({ message: 'Event successfully deleted' }) : res.status(500).send({ message: 'Wrong id' });
-        })
-        .catch(err => {
-            res.status(500).send({ message: err.message });
-        });
+    if(req.user.role === 1 || req.user.role === 2) {
+        events.destroy({where: {id: req.params.id}})
+            .then(event => {
+                event ? res.status(200).send({message: 'Event successfully deleted'}) : res.status(500).send({message: 'Wrong id'});
+            })
+            .catch(err => {
+                res.status(500).send({message: err.message});
+            });
+    } else res.status(500).send({ message: 'You have no rights' });
 });
 
 module.exports = routes;
